net: When using an external bond, persist its mac address
Bond devices may be created outside of VDSM and used by a network.
In such cases, the bond is "acquired" by VDSM.
The bond mac address is chosen based on the original creation tool/env.

When VDSM takes over, the bond mac address may change, leaving the bond
with a different mac address and thus potentially getting a different
dynamic IP address.

This patch preserves the mac address in such scenarios, attempting to
avoid cases in which the bond mac address change when acquiring it.
Upon the canonicalize stage, the bond type is detected and if it fits
the criteria, its mac address (hwaddr) is preserved.

Change-Id: I07dc3480c651ace01adf1926644cc2579faf0159
Signed-off-by: Edward Haas <edwardh@redhat.com>