Remove last usage of non-standard deathSignal
Replace usage of non-standard cpopen option "deathSignal" with
utils.terminating() helper.

This change also fix possible reaping of the dd child process if the
operation failed. We used to kill the process without waiting for it.

We are still using commands.AsyncProc which is not compatible with
python 3, more work is needed to replace it.

Change-Id: I1202af7bb2e7c6a429cd412f83427d284d451167
Signed-off-by: Nir Soffer <nsoffer@redhat.com>