net: Bond objectivize should attempt to use setup bond attrs
During the network objectivize step, if the network is based on a bond,
the bond is objectize with values read from netinfo.

With this patch, the bond objectivize is performed with bond attributes
defined in the requested setup.
If the bond was specified in the current setup transaction, its
attributes are used directly, otherwise, the bond attributes are fetched
from the running config (at canonicalize step).

Change-Id: I18c18e44edce72e7db8a84a909cc60faa531d4c3
Signed-off-by: Edward Haas <edwardh@redhat.com>