live-merge: Handle LookupError when tracking active layer merge job
When tracking jobs in queryBlockJobs, in case of active layer merge
pivot might already be completed, but the thread isn't marked as
succeeded yet. In this case, a LookupError is raised causing live merge
operation to fail. This isn't the desired behavior as in the next block
jobs query, the thread will probably be marked as succeeded and the flow
is supposed to continue as expected. In this patch, we correctly handle
this use case, by handling the exception and continue tracking the next
block job. In addition, we don't abort querying drive block jobs if the
drive is gone, a use case that is less likely to happen.

Change-Id: Ia2c172211740ed6d58a31b33cbf6778cac522a22
Bug-Url: https://bugzilla.redhat.com/1488542
Signed-off-by: Ala Hino <ahino@redhat.com>