virt: Make sure Drive.path is not None for file devices
Domain XML stored by libvirt and later delivered via _srcDomXML parameter on
migrations may omit `file' attribute from `source' element of CD-ROM and
similar devices if the file is an empty string.  When we attempt to read the
attribute and put it to DOM structures, we must ensure the value is not None,
otherwise DOM export will crash.

This patch assigns empty string to `path' parameter when `file' attribute is
missing in `source' element.

Change-Id: I8f30fae69d0cd2c790b0669357e069b8db9e7ace
Bug-Url: https://bugzilla.redhat.com/1507884
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>