net: Adjust dhclient detection for IPv6 requests
VDSM determines if a dhcp request is enabled on a network/device by
parsing the dhclient arguments.

It turns out that when the ifcfg initscripts executes dhclient for
an IPv6 request, it appends the host name (-H option) at the tail of the
arguments.
VDSM has assumed up until this patch that the last argument is the
device name and not another option. This patch attempts to detect the
device by ignoring any options at the tail.

Note: This solution does not cover cases where a flag ('-') is
positioned just before the device.
Example: sbin/dhclient ... -6 eth0

Change-Id: If6665fdda8852e959e0602f4c6fdbb2b0ebb7af8
Signed-off-by: Edward Haas <edwardh@redhat.com>