gencerts: use yesterday as the valid date for the CA cert
Following the same pattern as certs for engine, use yesterday's
date as the starting date for certificate validity, in case the
local hardware clock is not UTC (or will otherwise be drift
corrected).

This was already done for vdsmcert, but the CA also needs it, or
it will not be able to verify the chain.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1455441

Change-Id: I2a96a43916f4d525d8b63ede9448329ef9f7202c
Signed-off-by: Ryan Barry <rbarry@redhat.com>