vdsm-client: update documentation
In order to debug communication errors between Engine and Vdsm,
it is useful to have an option to install vdsm-client package
and run Vdsm commands.
The only problem is that vdsm-client looks for its certificate under
/etc/pki/vdsm, which is missing on Engine host.

We can overcome this by creating symbolic links to engine cert and key
files.
An explanation was added to vdsm-client doc.

Change-Id: I94604a7657089b786d4c642c9d5bf820b716a929
Bug-Url: https://bugzilla.redhat.com/1508375
Signed-off-by: Irit Goihman <igoihman@redhat.com>