vm: recreate disk device configuration from XML
Engine wants to send only the domain XML.
Unfortunately, we discovered some gaps in the implementation
of some storage operations (live merge, and thus live storage migration),
which are still heavily using the vm.conf data format.

To move forward, we rebuild the vm.conf data from the engine XML,
to present a sane environment to live merge.

This patch is safe because we
1. use anyway engine XML as source of truth, so we are duplicating
   the same information
2. we persist the data in the Engine XML

more patches are in the pipeline to make the storage operations
stop use the vm.conf data, so we can remove this one in the near
future.

Change-Id: I58fb9c5ab9fc6b779f6046c749c4ab61897ad345
Signed-off-by: Francesco Romani <fromani@redhat.com>