virt: Don't report a failed migration as running
When a migration fails, SourceThread._recovery method gets called.  VM status
is changed to UP at the end of the method and the corresponding status event is
sent.  However VM._getVmStatus checks for migration and SourceThread.migrating
still reports True, so "Migration source" is still sent as the VM status.

Let's fix that and make SourceThread.migrating return False when VM status is
set to UP.

Change-Id: Id9b5eb46af3ca391b467fbdae6e5155d52dc7727
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>