Revert "virt: Don't call destroy on destination after failed migration"
This reverts commit 3dffa7c2fd041d719f94037121a91da20107c464.

We removed the destroy call in the hope that libvirt and Engine always do the
right thing and by removing the call we can be safer from various problems.
However libvirt doesn't signal failed migrations on the destination and when a
migration fails quickly then Engine doesn't call destroy on the destination.
So although libvirt makes proper cleanup on the destination, Vdsm doesn't.  The
result is that Vdsm thinks the VM is still present on the destination after
failed migration, until its long wait for migration completion times out.

The simplest mitigation for now is to keep the destroy call from the source
host.

Change-Id: I7255b086aa627e5de04b992afb29af59f8a4f1c9
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>