net: netupgrade - Ignore unowned bonds and do not persist them
During a network upgrade flow where the system is in
unified persistence mode, external bonds have been automatically
acquired and persisted in the unified config.
This is not the desired behavior, bonds that are not owned by VDSM
should not be acquired.

This patch fixes this behavior and ignores bonds which are not owned.

Change-Id: I9f669c34daf0dd6089b7f85dfc60d06f60183228
Bug-Url: https://bugzilla.redhat.com/1422430
Signed-off-by: Edward Haas <edwardh@redhat.com>