net: Add Link Aggregation and MTU TLV
The lldpad driver is extended to support the
Link Aggregation TLV and the MTU TLV.

The Link Aggregation TLV might be marked as OUI 80c2,
subtype 7, or OUI 120f subtype 3. lldptool shows both
in the same way, so it is not possible to conclude
from lldptool's output to the OUI and subtype.

Change-Id: Ie7244886ae46c843d221a4c2976e370cf4cd23b0
Signed-off-by: Dominik Holler <dholler@redhat.com>