net: Rename the bond used for option mapping
In order to create the default bond options map, on every boot VDSM is
running a scan on all bond modes and detects the existing attributes and
their default values. It does so using a bond device with a random name
which has an unwanted side effect of logging the scan.

In order to explain these logs and have them marked as related to this
bond scan, the random generated bond name has been assigned a constant
prefix: "bondscan-"

Change-Id: Id418fca9b717a457379dff80bfd02c10574cfe08
Bug-Url: https://bugzilla.redhat.com/1495905
Signed-off-by: Edward Haas <edwardh@redhat.com>