vm: set device metadata in updateDriveParameters
When a VM method changes the parameters of one drive, we call the
the Vm.updateDriveParameters() method to persist the changes to both
the Drive object and to enable the recovery.

Up until oVirt 4.1, that meant to update the drive configuration dict
in vm.conf, and write a new recovery file.
Now we need to persist the relevant fields in the device metadata,
and synchronize the metadata with the libvirt Domain object.

Change-Id: I81f5d71cc3f1ac7a46cd7f06ab7d795a0762307f
Signed-off-by: Francesco Romani <fromani@redhat.com>