net: Logging failure to read the speed at debug level
Logging failure to read the speed has been done in WARN level, which
reached the journal, over spamming it.

It is enough to log it in debug level for now (not reaching the journal).
Even when the logging is missing, a value of 0 implies a possible
failure, therefore, the information lost has small impact.

Change-Id: I0afe60aa69a7a0d94ddbafe439e76ae5fe437824
Bug-Url: https://bugzilla.redhat.com/1510443
Signed-off-by: Edward Haas <edwardh@redhat.com>