vm: livemerge: generalize getVolumeInfo helper
The Vm._syncVolumeChain method defines a private getVolumeInfo helper.
The name is a little misleading: what the helper do is to find the
volume chain node matching a given volumeID; that node will be used
as up to date volumeInfo for the drive subject to live merge.

This patch generalizes the helper and makes it a free function.
We want to reuse it in a future patch, to make sure the livemerge flow
uses the drive object attributes - not its configuration stored in
vm.conf, so the latter can be eventually removed.

Change-Id: Id16492e9010676a7e9519a891cd0202c1191938e
Signed-off-by: Francesco Romani <fromani@redhat.com>