vm: livemerge: don't share volInfo
The Vm._syncVolumeChain method used to share some nested data
structure (volumeChain, volumeInfo) between the drive object
and the vm.conf dictionary.

This means that both places will hold a reference to the same, say,
list. However, we want to stop populating vm.conf - as soon as Engine
doesn't use it anymore -, so we need to start making sure that the
drive object, and not its conf, is the source of truth.

We start this process splitting the update of volumeInfo in two
parallel tracks, one for the drive object and one for its conf.
Nothing should change in terms of behaviour, because the data
is supposed to be the same.

To make sure, we add a safeguard and warning.

Change-Id: Id163c928e2910a3405da6a20c04e91e952f93380
Signed-off-by: Francesco Romani <fromani@redhat.com>