vm: doc: unset threshold when live merge completes
The last step of a live merge is a pivot. If this step
succeeds, we need to set a new threshold on the new active
layer. We can't do this immediately, because we can't make
any safe assumptions about the drive state.

So, the best thing to do is to treat the drive like a new one,
set its threshold_state to UNSET and let the drive monitor do
any needed check in the next monitoring cycle.

The drive monitoring is already re-enabled at the end of live merge.

This is a doc-only change. All the needed steps are already
done by the existing code.

Documentation-Only: yes
Change-Id: Icc7e8af30de6870542bbc1a59626ef98de85a2f2
Signed-off-by: Francesco Romani <fromani@redhat.com>