iscsiadm: Replace execCmd(async=True) with commands.start()
Replace usage of commands.execCmd and utils.AsyncProcessOperation with
commands.start().

Merge the unused session_rescan() and session_rescan_async() into
session_rescan(timeout). We raise now new IscsiSessionRescanTimeout if
the operation timed out, so the caller can handle the timeout.

This change revealed a serious bug in the old code: all errors during
scan were swallowed, since the only way to get the error is to call
AsyncProcessOperation.result(), which is not possible when you want to
timeout the operation, since it blocks until the process terminates.
With the new code, all scanning errors are reported to the caller. We
log now scanning successful termination, timeout and errors in the same
way fc-scan operation is logged.

Here is an example message when scanning times out:

    Scan failed: Timeout scanning iSCSI sesions (pid=1234, timeout=30)

Remove the now unneeded password hiding code in iscsiadm._runCmd, and
the additional hideValue parameters from the commands that need to hide
passwords. Instead, in the iscsi code calling these methods, do not
unprotect password sent to iscsiadm functions. The passwords are
unprotected just before running the function by commands.start().

Since we don't use AsyncProc now, the iscsi test moudle is compatible
now with python 3.

Change-Id: I4398df7127c722c84cee53917bce813f234e17c3
Signed-off-by: Nir Soffer <nsoffer@redhat.com>