vm: split updateDriveVolume,  updateDriveParams
Vm.updateDriveParams used to call updateDriveVolume implicitely.
Now we make the call explicit in the calling sites of updateDriveParams:
a future patch wants to handle the failure of updateDriveVolume, and
update the block thresholds only if the update succeeds.

Vm.updateDriveParameters used to call Vm._sync_metadata() right after
the internal call to Vm.updateDriveVolume.
There is not a dependency to those calls. Vm._sync_metadata() cares
about a subset of the drive parameters that
a) must be taken from the metadata, because are oVirt-specific and not
   represented, or can be safely inferred, from the domain XML
b) cannot be learned by Vdsm on its own

The up to date disk sizes, returned by Vm.updateDriveVolume, belong
to neither of those categories.

Change-Id: I9915da6442c02af86cb2a56a11ab80e056f8b62b
Signed-off-by: Francesco Romani <fromani@redhat.com>