virt: Set correct agent channel name when not present in conf
Engine used to send oVirt guest agent name in conf and/or in metadata.  That is
no longer true and the agent name is present just in the corresponding
<channel> element of the VM's domain XML.  The current code sets the legacy
channel name in such a case, which is wrong.  We should use the new agent
channel name instead.

We introduce a helper method to get the correct channel name.  If Engine
explicitly sends the channel name in conf, it's used.  Otherwise we must
distinguish between old Engine and new Engine.  New Engine sets the name in the
domain XML, so we look if it is present there.  If it is then it is used,
otherwise the legacy name is used.

Since oVirt guest agent is used in only three places, making the socket file
path, constructing domain XML and making compatibility symlinks, two of them
being legacy code, there's no strong need to store it in an attribute.  So we
make the new helper property replacing the original attribute.

Change-Id: I4757cce9784f06a1936465cd5ebdf6ac14c8db30
Bug-Url: https://bugzilla.redhat.com/1512572
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>