net: don't persist nameservers when using dhcp
Curently, when setting up a management network, we persist nameservers.
The problem is that we persist them even when DHCP is to be used.
Due to this, change of nameservers on DHCPd is not propagated to VDSM
unified persistent config. On boot, it will overwrite the older ones.

With this patch we don't persist nameservers when DHCP is used.

Change-Id: I099c9495b33312a4dce2cdc176982f6ed2876170
Signed-off-by: Petr Horacek <phoracek@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1508397