virt: Ignore disconnected domains in Vm._undefine_domain
Disconnected domains can't be undefined by definition, so it's pointless to
crash in the undefine call on domain cleanup.  A typical case where that
happens is when canceling an incoming migration – the destroy call fails on the
domain undefine call and the failure is reported to the user.  Not calling
undefine in that case is not a problem, since libvirt takes care of cleaning up
the migration destination VM.

So let's ignore _undefine_domain errors on Disconnected domains.

Change-Id: Ib2096c5081124d40996185311e1a7f184de14886
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>