net: save ovn-bridge-mappings on upgrade
OVS property ovn-bridge-mappings is used by ovn to connect its logical
networks to physical network/bridge. Support for this property was added
to VDSM only recently.

In case a user deployed networks on his hosts before this feature, and only
then upgraded to latest VDSM, mappings will be missing and provider physical
networks won't work.

With this patch we make sure that mappings are saved on upgrade.

Change-Id: I38cdf55c5a8bd65117d305abe1b50f495c42c336
Signed-off-by: Petr Horacek <phoracek@redhat.com>