vm: Extract a _update_drive_volume_size method
After drive's volume is extended, we have to update the drive size and
set a new threshold. Extract the code from __afterVolumeExtension to a
new method, since we need to do the same in __afterReplicaExtension, if
while extending the replica, we pivoted to the replica.

This code is covered by the current drive extension tests.

Change-Id: I128bb852aa5f9092d67e4ac4c24b6f4ce09e4753
Signed-off-by: Nir Soffer <nsoffer@redhat.com>