tests: Add replication extension tests
Extending drives during replication is complicated the tricky, and we
have no tests for this flow. Some extension tests cover this as a side
effect of checking the outcome. Add new tests for checking this flow.

Unfinished - added class dosctring with the flows we need to test, no
tests added yet.

Change-Id: Ieac5e2893aacd95aa9f8b86d77ec16ead5a5947f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>