hooks: The PYTHONPATH for hooking should not depend on hooks
The hooks module is used by VDSM to access the existing hooks on the
system. The hooks themselves are importing the hooking module which
exposes VDSM functionality to the hooks.

Current code is assuming a dependency between the hooks module and the
hook package, including the modules locations.
This patch removes the location dependency (which assumes that hooks and
hook are located in the same path).

Change-Id: I7e853df592b3b19695a082b12308dc55a2d3c2c3
Signed-off-by: Edward Haas <edwardh@redhat.com>