libvirtconnection: Relocate libvirt_password
libvirt_password is being used by libvirtconnection, however it is
defined under tool.configurators.
It seems more appropriate for the tool configurators to depend inwards
to VDSM.

This patch relocates libvirt_password and its constants to
libvirtconnection, allowing it to be solely dependent on standard libs
and common.
The following patch will relocate libvirtconnection to common.

Change-Id: I6ad107ec4a64a54f361c270db527fee6588547b8
Signed-off-by: Edward Haas <edwardh@redhat.com>