vm: get vmType from domain, not from conf
The commit 66971d3a8 stop adding defaults to vm.conf in the XML flow.
Unfortunately, we didn't notice that there is one place in the
Vm.getStats() flow on which the code assumes vm.conf['vmType']
to be present - this is no longer the case.

This patch solves this issue returning the vm_type from the
domain descriptor, not from the vm.conf: this is the right
thing to do in both flows.

Change-Id: Id5f37b60924fccea90525597d3a2cdf08257dd5d
Signed-off-by: Francesco Romani <fromani@redhat.com>