virt: make sure to store device metadata
When a VM is started from domain XML, Engine is in charge
to set the proper device metadata, so Vdsm just needs to
keep them up to date, should any action change them.

However, Vdsm must support the old vm.conf initialization
method, and, given that recovery files are long gone, it
is now in charge to save -and not just update- the per-device
metadata in this flow.

This patch implements the code that saves that metadata.
Without this patch, Vdsm works fine until it is restarted;
after the restart, random action could fail with apparently
unrelated exceptions, most likely AttributeError.

Change-Id: I9a87fcf9c65062f67b74be1554bbc4ca92347e0a
Bug-Url: https://bugzilla.redhat.com/1516660
Signed-off-by: Francesco Romani <fromani@redhat.com>