automation: Fix fc27 build when running on fc26 slave
We used to install pip using easy_install, because pip was not available
on the EL7 slaves. However in recent Fedora, easy_install is not
available, and the build will fail with:

00:01:15.007 + easy_install pip
00:01:15.008 ./automation/check-patch.sh: line 53: easy_install: command not found

python-pip is available on EPEL, which is ok to use in oVirt CI.
Require python-pip and remove easy_install and the python-setuptools
required for it.

Example failing build:
http://jenkins.ovirt.org/job/vdsm_master_check-patch-fc27-x86_64/100/console

Change-Id: I33bfaf505445345d9bfcc5ba4f5ba221d5f9578f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>