virt: don't validate params on incoming migration
When Vdsm receives one incoming migration of a VM
created with domain XML, it will incorrectly try to
validate the params. But they will be missed, so
the validation fail. The correct thing to do is to
skip validation in this case, but we must distinguish
further between VMs created in 4.2 with domain XML
or VMs created in 4.1 with vm.conf.

In the latter case, we should keep validating the parameters.

Change-Id: Ia90ab60e1386fc5674c4e9adc0e61a8ea681c2eb
Bug-Url: https://bugzilla.redhat.com/1516405
Signed-off-by: Francesco Romani <fromani@redhat.com>