storage: Add mpath health monitor
Adding the mpathhealth module.
The mpathhealth.Monitor registers to udev.MultipathListener
in order to be updated on MultipathEvents.

The Monitor maintains a data structure containing the
faulty paths and the number of valid paths for each device.

The status of the devices will be first read in the 'start'
method, ensuring the integrity before receiving events.

The data is available through the 'status' method.

Here an example of the data:
{
    "uuid-2": {
        "valid_paths": 1,
        "failed_paths": [
            "sdba",
            "sdbb"
        ]
    },
    "uuid-1": {
        "valid_paths": 0,
        "failed_paths": [
            "sdaa",
            "sdab"
        ]
    }
}

The 'Mpath health monitor' will be used as part of the multipath event
feature:
https://www.ovirt.org/develop/release-management/
features/storage/multipath-events/

Change-Id: Ieb02d7c637cb42cb5e0f55ff653c3bf3f662cc0e
Signed-off-by: Fred Rolland <frolland@redhat.com>