udev: Support monitor hot[un]plug
We don't need to support this yet, as we plan to register all monitor
during startup, but we have tests using this for testing registration
from a callback, and implementing this is not much more complicated then
blocking this.

Hot-plug flow:

1. Listener add the monitor to the monitors set - at this point the
   monitor can receive events from the observer thread
2. Listener start the monitor for checking current system state
3. If the monitor failed to start, it is removed from the monitors set

Hot-unplug flow:
1. Listener remove monitor from monitors set
2. Listener stops the monitor

It is harder to write a monitor that can be hot-plugged, since you must
handle events from the observer thread while you check the current
system state, so this is usage is not recommend.

Change-Id: Ibdb737749f930799934346e0b885c7ee55ebc220
Signed-off-by: Nir Soffer <nsoffer@redhat.com>