lvmconf: Introduce the storage lvmconf module
This module helps to access LVM configuration, required for analyzing
and modifying current configuration when configuring a host. We also
plan to use it later for verifying that a host has a safe LVM
configuration during Vdsm startup.

We use the amazing augeas library, but provide an easier to use
ConfigParser like interface.

Here is example usage of this module:

    with lvmconf.LVMConfig() as conf:
        old_filter = conf.getlist("devices", "filter")
        ...
        conf.setlist("devices", "filter", new_filter)
        conf.save()

Change-Id: I31f088df483d75aa3342b5d03660a997ffbbfd93
Bug-Url: https://bugzilla.redhat.com/1523152
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Germano Veit Michel <germano@redhat.com>