vm: avoid NotConnectedError on shutdown
The API getAllIoTunePolicies() may let the NotConnectedError
bubble up if we hit a race on shutdown.
The error is benign - no real harm is done, yet it adds noise
in the logs, so this patch adds a new method to deal with
it in a more grceful way.

This patch is also the first step to clean up the getIoTune*()
mess we currently have in the Vm class.

Change-Id: I7364d6d19488ee58b5f989e060b33d0b492a14d0
Related-To: https://bugzilla.redhat.com/1503468
Signed-off-by: Francesco Romani <fromani@redhat.com>