mpathhealth: Implement Monitor start method
Implementing mpathhealth.Monitor.start method.
This method reads the status of the multipath devices, collecting the
failed paths and counting the active ones.
If all the paths are active, no status is kept.

This is needed to collect the faulty devices before starting to listen
to the udev events.
Also report to the API the actual state even if the device paths were
faulty before the Vdsm started.

Change-Id: I638ea6187231b54da44ebba2db9065851fa4868c
Signed-off-by: Fred Rolland <frolland@redhat.com>