virt: metadata: correctly store portMirroring
The portMirroring attribute of the network devices is one of the
few supported metadata items that is a sequence.

While the load from metadata worked fine, the store had a bug:
the top level element ('portMirroring') was lost.

Instead of

  <device mac_address='...'>
    <network>foo</network>
    <portMirroring>
      <network>bar</network>
    </portMirroring>
  </device>

The code incorrectly stored

  <device mac_address='...'>
    <network>foo</network>
    <network>bar</network>
  </device>

This patch fixes the issue and adds a test to exercise the code,
to make it unlikely to break again.

Change-Id: Ib85e5f5cb027455d9f752f4886eccc05b089c098
Bug-Url: https://bugzilla.redhat.com/1496719
Signed-off-by: Francesco Romani <fromani@redhat.com>