lvmfilter: Analyze current LVM filter
The current version of vdsm-tool config-lvm-filter is too naive and not
helpful enough. It will always recommend to configure a filter, even if
the host is already configured. When configuration is needed, it asks
the user to configure, instead of writing the configuration.

The goal of this work is to have automatic LVM filter configuration, as
part of the host deployment. We also need to check the current
configuration when Vdsm starts, and fail if LVM filter needs
configuration.

To allow this behaviour, the lvmfilter module provides now the analyze()
function. This function accept the current lvm filter, and the wanted
lvm filter, and return an Advice object suggesting how to proceed.

Bellow are example outputs for typical use cases.

Host is not configured:

    >>> lvmfilter.analyze(None, ["a|^/dev/sda2$|", "r|.*|"])
    Advice(action='configure', filter=["a|^/dev/sda2$|", "r|.*|"])

Host is configured:

    >>> curent_filter = wanted_filter = ["a|^/dev/sda2$|", "r|.*|"]
    >>> lvmfilter.analyze(current_filter, wanted_filter)
    Advice(action='unneeded', filter=None)

Host is configured using different syntax:

    >>> current_filter = ["a/^/dev/sda2$/", "r\.*\"]
    >>> wanted_filter = ["a|^/dev/sda2$|", "r|.*|"]
    >>> lvmfilter.analyze(current_filter, wanted_filter)
    Advice(action='unneeded', filter=None)

Host is configured but filter items have different order:

    >>> curent_filter = ["a|^/dev/sda2$|", "a|^/dev/sdb2$|", "r|.*|"]
    >>> wanted_filter = ["a|^/dev/sdb2$|", "a|^/dev/sda2$|", "r|.*|"]
    >>> lvmfilter.analyze(current_filter, wanted_filter)
    Advice(action='unneeded', filter=None)

Host is configured with a different LVM filter. In this case we
recommend to change the filter, but we will not confiugre it
automatically:

    >>> curent_filter = ["a|^/dev/sda2$|", "a|^/dev/sdb2$|", "r|.*|"]
    >>> wanted_filter = ["a|^/dev/sdb2$|", "r|.*|"]
    >>> lvmfilter.analyze(current_filter, wanted_filter)
    Advice(action='recommend', filter=["a|^/dev/sdb2$|", "r|.*|"])

Change-Id: I18260b2a7fe01b1e0e46e853e9ded61d03209709
Bug-Url: https://bugzilla.redhat.com/1449968
Signed-off-by: Nir Soffer <nsoffer@redhat.com>