virt: drivemonitor: test for negative threshold
We discovered that the drivemonitor can set negative threshold
if the drive size is unexpectedly too small.
It is first a storage issue if the disk is created below the minimum
supported size (1 GiB), but the drivemonitor should nevertheless be
more robust and not set impossible values.

This patch adds a test to demonstrate this issue.

Change-Id: I92bae8da5a08f125fdf32c7f94941e3c9d9d70aa
Signed-off-by: Francesco Romani <fromani@redhat.com>