virt: drivemonitor: avoid negative thresholds
We discovered that the drivemonitor can set negative threshold
if the drive size is unexpectedly too small.
It is first a storage issue if the disk is created below the minimum
supported size (1 GiB), but the drivemonitor should nevertheless be
more robust and not set impossible values.

This patch makes sure that the drive monitor never sets
thresholds below a minimum (constant value) of 1 MiB.

Change-Id: Ie150959976689e2f9a0347d00980d2fb67986de9
Signed-off-by: Francesco Romani <fromani@redhat.com>