automation: Add timeout to the tests run in check-merged
Several occurrences have been seen where the lago shell which runs the
tests got stuck and never returned. It caused jenkins to cancel the job
by sending a termination signal after 3 hours.
Unfortunately, when doing so, the logs are not collected and processes
are not correctly cleaned up from the worker machine.

A timeout is added to the execution of the lago shell which runs the
tests, limiting it to 1 hour.

Change-Id: I34734015dc3fc58a2a39884795a47dfe302ef3d4
Signed-off-by: Edward Haas <edwardh@redhat.com>