virt net: Support setting a display network on a non-vm net
A display network that is set on a non-vm network caused VM/s to fail
starting. The libvirt network created for the display network was
wrongly assumed to be based on a bride.

The libvirt network created now for the display is now supporting both
bridged and bridgeless host networks.

Change-Id: I9aa7d7910f0ccc16c21ba48f337b6026d37b98a1
Bug-Url: https://bugzilla.redhat.com/1518688
Signed-off-by: Edward Haas <edwardh@redhat.com>