xlease: Simplify DirectFile interface
Previously write did not fsync, and the caller could fsync if needed
using DirectFile.fileno(). We want to add an implementation that will do
the I/O outside of Vdsm, so we cannot assume an instance has a file
descriptor.

The use pattern of DirectFile is:
- write single block to storage for atomic changes like allocating a
  lease or marking an index as updating.
- write multiple blocks when formatting an index

In all cases, we like to fsync after every write. We don't have a use
case for write without fsync, so having a generic file like interface
is not needed.

This patch removes the fileno() interface, and fsync on every write.

Change-Id: I52801f778823d8063b3ca2d1ac3731ac363c5e38
Bug-Url: https://bugzilla.redhat.com/1527827
Signed-off-by: Nir Soffer <nsoffer@redhat.com>