xlease: Add InterruptibleDirectFile class
InterruptibleDirectFile implements the DirectFile interface using dd for
reading and writing from storage, and ioprocess for checking file size.

If storage becomes non-responsive, the thread calling
InterruptibleDirectFile will block, but the it will remain
interruptible, so it can be restarted if needed. In this case systemd
will take care of the stuck child process.

The new implementation will be used for file storage domain, when
storage may block for infinite time. On block storage our multipath
configuration limits the maximum block time, so we may block for 15-30
seconds on the first time all paths become failed. After that any access
to storage will fail immediately. This is also the existing convention
in vdsm; we access storage directly when using block storage, and via
external process for file storage.

4.1 notes:

- The cmdutils, commands and constants imports were changed to import
  from ``vdsm`` instead of ``vdsm.common``.

Change-Id: Ib05c706c0b93be81098eac5ad2b1af7cb6f03ec9
Bug-Url: https://bugzilla.redhat.com/1527827
Signed-off-by: Nir Soffer <nsoffer@redhat.com>