vmxml: reimplement ElementTree.tostring()
In Python3, ElementTree has become more picky, and
emits the xml header (aka the declaration):

  <?xml encoding=... >

in even fewer circumnstances than it used to do in python2.
With the switch of python3, Vdsm may silently start to lose
those declarations.

The actual impact should be negligible, but ElementTree
provides a simple fix with the `xml_declaration` argument
of the write() method.

Unfortunately, if we use the tostring() utility, we don't
have easy access to that argument.
The simplest way to fix this is to reimplement tostring()
locally, borrowing a good chunk of the Python3 implementation.

This is actually way less scary than it sounds, since the
implementation is just one handful lines long.
This way we can always have the XML declaration both in python2
and in python3.

Change-Id: I4944372c8d1d96f33901ceb9f1606309b99705be
Signed-off-by: Francesco Romani <fromani@redhat.com>