fileSD: Use InterruptibleDirectFile backend
FileStorgeDomainManifest uses now InterruptibleDirectFile for the
external leases backend.  This ensure that the process does not become
uninterruptible if storage is non-responsive.

This change bypass the python bug[1] causing the entire process to hang
when opening the special xleases volume. Now the operation is done in
the child process outside of Vdsm.

[1] https://bugzilla.redhat.com/1520068

Change-Id: I8a0eb42baa73b6c0a877ad9acbe53ac51f31bfea
Bug-Url: https://bugzilla.redhat.com/1527827
Signed-off-by: Nir Soffer <nsoffer@redhat.com>