xlease: Fix handling of truncated volume
Trying to read a truncated external leases volume was stuck in endless
loop because EOF was not handled. This is not an issue in practice,
since having a truncated xleases volume means you have much bigger
trouble, but it should be fixed anyway.

This error will raise now TruncatedIndex error:

    TruncatedIndex: Couuld not read index, expected 256512 bytes,
    available 256000 bytes

Change-Id: I5bcc53d0218cfc76e70a43d0ef38d46e226e65b9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>