lvmfilter: Filter out the master LV on the SPM
When running on the SPM when the master domain is a block storage
domain, the master LV is mounted. This cause two issues:

1. The tool suggest to include the master LV devices in the filter. This
   is harmless since the master LV is never used by VMs, but a bad idea in
   general. We don't want the master LV to be available on a host.

2. If a filter was already configured, which does not include the master
   LV devices, the tool fail to check the master VG info, since the
   devices are hidden by the filter. The ``lvs`` command fail with error
   5, and we fail the configuration.

Additionally, if a host has an incorrect filter hiding some devices
needed by the host, the tool may fail to create a filter, since the
``vgs`` command will fail reading partial VG. This is very unlikely
since such host may not boot, but easy to avoid.

Fix both problems by:

1. Override the current filter with permissive filter ["a|.*|"] when
   looking up VG info in mounted LVs. This ensure that ``lvs`` and
   ``vgs`` commands succeed, and we can create a correct filter
   regardless of the currently configured filter.

2. Read the VG tags of mounted LVs, and skip oVirt LVs using the special
   ``RHAT_storage_domain`` VG tag.

When running in verbose mode (-vvv), we will see now this debug message
about the master LV:

    MainThread::DEBUG::2017-12-06 03:52:31,300::lvmfilter::134::lvmfilter::(find_lvm_mounts)
    Skipping oVirt logical volume u'/dev/mapper/6ffbc483--0031--403a--819b--3bb2f0f8de0a-master'

Change-Id: I74309fcc3960e7988b2baf0569b89fd465ebfafe
Bug-Url: https://bugzilla.redhat.com/1523152
Signed-off-by: Nir Soffer <nsoffer@redhat.com>