localdisk: Support host LVM filter
The ovirt-local VG devices are not not needed by the host, so they are
not included in the LVM filter. If the user created LVM filter with
vdsm-tool config-lvm-filter, the localdisk helper process will fail to
access the ovirt-local VG.

Fixed by adding a permissive filter in localdisk-helper LVM operations.

Change-Id: Ib4acd7a5f7e55fc7d07bac0d67f983d82e213524
Bug-Url: https://bugzilla.redhat.com/1523232
Signed-off-by: Nir Soffer <nsoffer@redhat.com>