virt: Make sure XML configuration is present on resume from snapshot
XML Engine stopped sending `xml' parameter when resuming from a live snapshot
with memory.  We use _srcDomXML parameter to initialize the domain, but `xml'
parameter is missing in self.conf.  Then Vdsm thinks that we are running with
legacy conf, which is actually not present too.  The result are errors when
trying to start the VM.

This patch fixes that by copying _srcDomXML to self.conf['xml'].

Change-Id: Ifc8f8f59401021b6a847f74aa0b03460fdb2c47c
Bug-Url: https://bugzilla.redhat.com/1523399
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>