storage: Add missing test for unset diskType
Since ovirt 4.2 we always have a disk type, but in older versions
diskType was not sent from engine, and was discovered by vdsm, assuming
that the type can be file or block. Since we added support for network
disk type, all the tests were converted to use explicit disk type, the
legacy case untested.

Change-Id: Ifd958d709303326dfa0f8f099a0d0f173c213301
Signed-off-by: Nir Soffer <nsoffer@redhat.com>