tests: Add the missing needs_monitoring tests
We just fixed a deadlocked in Drive.needs_monitoring. If we had tests
for this code testing the forgotten no diskType use case, we could
discover the deadlock during the first build.

This code is being tested by the drive monitor tests, but these tests
are too big and complicated. Having simple unit tests for the Drive
class will make it easier to refactor it later.

Change-Id: I99f4d0556d63bf4c9a3227adf0ccd40bc35acd0b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>