storage: Set diskType early and safely
We suffer from lazy initialization of Drive.diskType since it was
introduced in http://gerrit.ovirt.org/4283. This makes testing harder,
and recently caused a deadlock during migration.

In 4.2 engine is always sending the diskType since it knows the type of
the storage, simplifying vdsm flows. Move diskType initialization to
prepareVolumePath, normalizing drive creation in legacy flows, and
simplifying Drive code.

We used to check if a path is a block device using os.stat(), which can
block for unlimited time on non-responsive NFS server. We can avoid this
now using the information returned by Image.prepare() for vdsm images.

Note that we use 'network' disk type only if engine specified this
value. Therefore if the value was not specified, gluster volume must use
'file' diskType.

For GUID format, we don't need to check anything, since GUID format
means we are using a multipath device.

For other types of volumes, we are not supporting "block" disk type so
we can default to "file" unless engine ask to use another type, like
"network" when using Ceph storage.

Change-Id: I2338ce0bac95f82ed56e97dea11b8a0d0f38174d
Bug-Url: https://bugzilla.redhat.com/1530072
Signed-off-by: Nir Soffer <nsoffer@redhat.com>