tox: Use latest pylint version
Our version is too old, and we fail to detect broken code because of
this. Lets try to use latest pylint.

I hoped that latest pylint fixed detection of missing attributes on
builtin types (https://github.com/PyCQA/pylint/issues/1736). This is
still broken even on master with python 2.7, but works with python 3.
We should start run pylint with both python 2.7 and python 3.

Change-Id: Ibc2d12ab29ef99a6b905408a42b75098a27e93f0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>