pylint: fix no value for argument error
pylint complains about "no value for argument 'action' in function
call" when calling private function _parse_args.
It seems that although the function declaration contains 'action' as
argument, no method is actually passing it directly when calling
_parse_args.
'action' argument was removed from the declaration and now it's part of
'args'.

Change-Id: I0ddabcecdc7ad6053c6912e78c842ea483ed4bf5
Signed-off-by: Irit Goihman <igoihman@redhat.com>